[iOS] Refactor: NOW 다시듣기, 매거진 섹션 코드 리팩토링
모델 분리와 뷰 간 분리
매거진 더미 이미지 추가
