+++ iOS/MiniVIBE/MiniVIBE/HomeView/HomeMagazineSection/HomeMagazineItem.swift
@@ -1,0 +1,14 @@
+//
+//  HomeMagazineItem.swift
+//  MiniVIBE
+//
+//  Created by GH Choi on 2020/11/24.
+//
+
+import Foundation
+
+struct HomeMagazineItem: Identifiable {
+    let id = UUID()
+    var imageURL: String
+    var description: String
+}
