+++ iOS/MiniVIBE/MiniVIBE/HomeView/HomeNowSection/HomeNowReplayItem.swift
@@ -1,0 +1,14 @@
+//
+//  HomeNowReplayItem.swift
+//  MiniVIBE
+//
+//  Created by GH Choi on 2020/11/24.
+//
+
+import SwiftUI
+
+struct HomeNowReplayItem: Identifiable {
+    var id = UUID()
+    var albumArt: Image
+    var description: String
+}
